Moved classes to a 'security' folder
