+++ University_Simulator/src/main/java/com/challenge/alkemy/security/MyUserDetails.java
@@ -1,4 +1,4 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.security;
 
 import java.util.ArrayList;
 import java.util.Collection;
