+++ University_Simulator/src/main/java/com/challenge/alkemy/security/WebSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.challenge.alkemy;
+package com.challenge.alkemy.security;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
